requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.25
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.10
python-dotenv==1.0.0
marshmallow==3.20.1
flask-apispec==0.11.4
flask-cors==4.0.0

# Development dependencies
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.3.2
flake8-pyproject==1.2.3
black==24.2.0
autoflake==2.3.0
isort==5.13.2
